+++ src/ocio/Deh2020.cpp
@@ -8,13 +8,13 @@ namespace dehancer::ocio::DEH2020 {
 
     namespace forward {
         extern float __lut__data__[];
-        extern uint  __lut__size__;
-        extern uint  __lut__channels__;
+        extern size_t  __lut__size__;
+        extern size_t  __lut__channels__;
     
         DHCR_LutParameters lut::params = {
                 .enabled = true,
-                .size = forward::__lut__size__,
-                .channels = forward::__lut__channels__,
+                .size = static_cast<uint>(forward::__lut__size__),
+                .channels = static_cast<uint>(forward::__lut__channels__),
                 .data = forward::__lut__data__,
         };
 
@@ -23,13 +23,13 @@ namespace dehancer::ocio::DEH2020 {
     namespace inverse {
 
         extern float __lut__data__[];
-        extern uint  __lut__size__;
-        extern uint  __lut__channels__;
+        extern size_t  __lut__size__;
+        extern size_t  __lut__channels__;
     
         DHCR_LutParameters lut::params = {
                 .enabled = true,
-                .size = __lut__size__,
-                .channels = __lut__channels__,
+                .size = static_cast<uint>(__lut__size__),
+                .channels = static_cast<uint>(__lut__channels__),
                 .data = __lut__data__
         };
     }
